refactor(mock): remove unused pUuid parameter from migration func
Removed the unused pUuid parameter from several migration-related
function declarations in the mock interface. The parameter was
commented out in the implementation but still present in the header
file, causing inconsistency. This change cleans up the interface and
removes dead code.

The affected functions are: - MigrationGetCert -
MigrationCheckPeerPk - MigrationGetVRootCipher -
MigrationImportVRootCipher - MigrationNotify